Moving assets around so pages can be rendered directly
This moves a bunch of js and css assets around so that pages can be rendered directly.
Was caused because if we load the HTML template directly, rails serves up public js/css files
as mime type of text/html, which the client will then not execute/render. Therefore, we want
the rails server to serve these assets when the client firsts requests a page. In order for that
to happen I needed to move the js/css to the rails assets folder (out of public) and include them
in the application template using rails tags. They also need to be included to be precompiled
for production