Adding player stats to team landing
This commit bundles a number of things together, but primarily adds
the ability to see player stats when you click on a player on the team roster.
It adds the stat table as a directive and tweaks the teams controller so that
the directive will have access to said data.

It also tweaks the routes, removing some unused static pages. Also updates the
API to accept XML endpoints a well as passing back the team name to the team roster
controller.